+++ lib/openproject-mattermost.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require "open_project/mattermost/version"
-require "open_project/mattermost/classifier"
 require "open_project/mattermost/formatter"
+require "open_project/mattermost/classifier"
 require "open_project/mattermost/client"
 require "open_project/mattermost/dispatcher"
 require "open_project/mattermost/notification_handler"